// src/pages/InquiryManage/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    customerName: string,
    pageNumber: number,
    pageSize: number,
    productName: string,
    queryStatus: number
  };

  type TableItem = {
    createTime: string,
    customerName: string,
    productsName: string,
    purchaseAmount: number,
    queryStatus: number,
    replyContent: string,
    replyUserName: string,
    saleStockOrderId: number,
    saleStockOrderNo: string,
    salesName: string
  }
}